// client/src/redux/constants/actionTypes.ts
export enum ActionType {
    AUTH = "AUTH",
    LOGOUT = "LOGOUT",
    UPDATE_USER = "UPDATE_USER",
    AUTH_ERROR = "AUTH_ERROR",
    CLEAR_AUTH_ERROR = "CLEAR_AUTH_ERROR",
}

// export const AUTH = "AUTH";
// export const LOGOUT = "LOGOUT";
// export const UPDATE_USER = "UPDATE_USER";

// export const AUTH_ERROR = "AUTH_ERROR";
// export const CLEAR_AUTH_ERROR = "CLEAR_AUTH_ERROR";